Move line item to separate file
